Introduces lazy initialization of Maps
Since we support stencil operation on subsets a good deal
of bookkeeping calculation is necessary. This was done for
all subset/map combinations at startup. This can, depending
on the local lattice size result in significant startup time.

The improvement uses two paths: 1) split the computation into
a part that has to be done at startup and defer the rest to a
later time (on-demand). 2) Evaluate on a per-subset basis
instead of for all subsets registered.

This should help startup times for large local problems.

(needs testing and cleanup)