Complete rewrite of sumMulti
The new implementation makes use of two dedicated jit
kernels where one of them uses the new multi1d support of
the cache. With this it was possible to relax all previous
restrictions such that sumMulti can be used on all Sets -
like with the original QDP++ version. Reductions are off-
loaded to the GPU and the jit kernels reduce on the thread
block while a single thread per block stores back to off-chip
memory.

The previous version made use of the existing reduction
kernels used for sum and friends. At the time this was
the way to go because of the missing multi1d support for
jit function arguments. However, to get this to work a
special 'strided' sitetable was needed - a hack/workaround
that introduced further restrictions down the line. E.g.
Sets created with time functions and executed with a machine
grid with time parallelized over several compute nodes would
not work with sumMulti. All these restrictions are gone now.