Remove unnecessary code in the reduction kernels
The reduction kernels contained code to find the smallest
power of 2 that is equal or larger to the thread block
size. It is not necessary to find this out this way since
these particular kernels already launch with a thread block
size of a power of 2 and this datum is available.