Mid-development multiXd jit support (II)
1) Adds a specialization for multi1d/2d/3d for QDP scalars.
At initialization memory is allocated from the device memory
cache in form of the newly added arrays. Subsequently the
array of scalars' resources are set to use the array. This allows
two views of the same memory: individual scalar and whole array.

This is not done yet. The new sumMulti is deactivated for now.

2) Adds the -poolmemset <unsigned> run-time option. It allows
to initialize the GPU device memory with a set value and is
probably only useful for debugging purposes.