Concludes the multiXd jit support for QDP scalar
Re-enables the new jit version of sumMulti. For this, the
'whole array' view of an cached array is needed - as opposed
to the scalar view through multiXd.

Also, since before there was no array view of a multiXd
of QDP scalars, the original globalSumArray needed to be
implemented by doing repeated call-outs to QMP sum. This
restriction is now relaxed and the array sum functionality
is now used for globalSumArray.