Memory pool bi-section
Choosing the right size of the device memory pool was always
a bit tricky and finding it systematically came necessarily
with lots of experimentation. The newly added memory pool
bisection routine tries to help finding the right memory
pool size within a single (production) run.

1) Setting the runtime option '-poolbisect' lets the program,
at the end of its runtime, perform a bisection search which
repeatedly bisects a min., max. memory interval in order to
find the maximal memory pool size that would still allow to
launch the device kernel with the most demand of free memory.
It does so by monitoring the jit kernels' usage of local memory
during the course of program execution. At the end of the
program's runtime the routine repeatedly tests different memory
pool configurations. The result of the bisection is given in
form of a runtime option to set the memory pool size, i.e.
'-poolsize <x>' and can be used to subsequent program
executions.

The '-poolbisect' option does not alter any parts of the
calculation. It is intended to be used in the first run once
the machine grid geometry has settled for a given problem
size. Still, this run is a production run. The wallclock time
overhead of the bisection is very small. Subsequent runs are
then launched without the bisect option but with the resulting
-poolsize option instead.
In principle each change of the local lattice volume requires
a new bisection run although the bisection result might change
only marginally.

2) The runtime option '-bisectmax <poolsize>' allows to limit
the upper limit of the bisection interval. This is probably
only useful when running multiple instances of the program
on the same CUDA device.

3) The memory cache has received a suspend()/resume() option
which lets the allocator temporarily release the big memory area
used as the memory pool.