Removed duplicate qdp_llvm.cc from list of sources
qdp_llvm.cc was listed in the source lists for both libqdp and libjit
I removed the one from libqdp since it links against libjit, and also
likely we want to compile it with C++-14.

This fixed an error (double free or corruption) on Summit at program exit
amongst the global destructors,  which arose from destroying the map  ptx_db::db when the -ptxdb option
was given. I am guessing the symbol was defined in both libjit and libqdp's qdp_llvm.o.
I am surprised that there was not a duplicate symbol issue at link time.

Best,
 B