Fix codegen for SubTypes on ROCm
Codegen happened on the primary node and other
nodes relied on the object file being present on
the filesystem. This is a problem on SubTypes when
the primary node has no work load and thus does not
generate code.

Solution: All nodes generate code (like is the case
with the CUDA backend). To avoid file clashes the
nodenumber was added to the filename, like

module_eval0_node_0.o

Also clean up files afterwards. Added new parameter
'-keep-files' with obvious effect.